UnitfulData

Documentation for UnitfulData.

UnitfulData.Bps

UnitfulData.Byte

UnitfulData.Data

UnitfulData.EBps

UnitfulData.EByte

UnitfulData.EHart

UnitfulData.Ebit

UnitfulData.Ebps

UnitfulData.Edibit

UnitfulData.EiBps

UnitfulData.EiByte

UnitfulData.Eibit

UnitfulData.Eibps

UnitfulData.Enat

UnitfulData.Enybl

UnitfulData.Esh

UnitfulData.Etribit

UnitfulData.Etrit

UnitfulData.GBps

UnitfulData.GByte

UnitfulData.GHart

UnitfulData.Gbit

UnitfulData.Gbps

UnitfulData.Gdibit

UnitfulData.GiBps

UnitfulData.GiByte

UnitfulData.Gibit

UnitfulData.Gibps

UnitfulData.Gnat

UnitfulData.Gnybl

UnitfulData.Gsh

UnitfulData.Gtribit

UnitfulData.Gtrit

UnitfulData.Hart

UnitfulData.KiBps

UnitfulData.KiByte

UnitfulData.Kibit

UnitfulData.Kibps

UnitfulData.MBps

UnitfulData.MByte

UnitfulData.MHart

UnitfulData.Mbit

UnitfulData.Mbps

UnitfulData.Mdibit

UnitfulData.MiBps

UnitfulData.MiByte

UnitfulData.Mibit

UnitfulData.Mibps

UnitfulData.Mnat

UnitfulData.Mnybl

UnitfulData.Msh

UnitfulData.Mtribit

UnitfulData.Mtrit

UnitfulData.PBps

UnitfulData.PByte

UnitfulData.PHart

UnitfulData.Pbit

UnitfulData.Pbps

UnitfulData.Pdibit

UnitfulData.PiBps

UnitfulData.PiByte

UnitfulData.Pibit

UnitfulData.Pibps

UnitfulData.Pnat

UnitfulData.Pnybl

UnitfulData.Psh

UnitfulData.Ptribit

UnitfulData.Ptrit

UnitfulData.TBps

UnitfulData.TByte

UnitfulData.THart

UnitfulData.Tbit

UnitfulData.Tbps

UnitfulData.Tdibit

UnitfulData.TiBps

UnitfulData.TiByte

UnitfulData.Tibit

UnitfulData.Tibps

UnitfulData.Tnat

UnitfulData.Tnybl

UnitfulData.Tsh

UnitfulData.Ttribit

UnitfulData.Ttrit

UnitfulData.YBps

UnitfulData.YByte

UnitfulData.YHart

UnitfulData.Ybit

UnitfulData.Ybps

UnitfulData.Ydibit

UnitfulData.YiBps

UnitfulData.YiByte

UnitfulData.Yibit

UnitfulData.Yibps

UnitfulData.Ynat

UnitfulData.Ynybl

UnitfulData.Ysh

UnitfulData.Ytribit

UnitfulData.Ytrit

UnitfulData.ZBps

UnitfulData.ZByte

UnitfulData.ZHart

UnitfulData.Zbit

UnitfulData.Zbps

UnitfulData.Zdibit

UnitfulData.ZiBps

UnitfulData.ZiByte

UnitfulData.Zibit

UnitfulData.Zibps

UnitfulData.Znat

UnitfulData.Znybl

UnitfulData.Zsh

UnitfulData.Ztribit

UnitfulData.Ztrit

UnitfulData.aHart

UnitfulData.adibit

UnitfulData.anat

UnitfulData.anybl

UnitfulData.ash

UnitfulData.atribit

UnitfulData.atrit

UnitfulData.bit

UnitfulData.bps

UnitfulData.cHart

UnitfulData.cdibit

UnitfulData.cnat

UnitfulData.cnybl

UnitfulData.csh

UnitfulData.ctribit

UnitfulData.ctrit

UnitfulData.dHart

UnitfulData.daHart

UnitfulData.dadibit

UnitfulData.danat

UnitfulData.danybl

UnitfulData.dash

UnitfulData.datribit

UnitfulData.datrit

UnitfulData.ddibit

UnitfulData.dibit

UnitfulData.dnat

UnitfulData.dnybl

UnitfulData.dsh

UnitfulData.dtribit

UnitfulData.dtrit

UnitfulData.fHart

UnitfulData.fdibit

UnitfulData.fnat

UnitfulData.fnybl

UnitfulData.fsh

UnitfulData.ftribit

UnitfulData.ftrit

UnitfulData.hHart

UnitfulData.hdibit

UnitfulData.hnat

UnitfulData.hnybl

UnitfulData.hsh

UnitfulData.htribit

UnitfulData.htrit

UnitfulData.kBps

UnitfulData.kByte

UnitfulData.kHart

UnitfulData.kbit

UnitfulData.kbps

UnitfulData.kdibit

UnitfulData.knat

UnitfulData.knybl

UnitfulData.ksh

UnitfulData.ktribit

UnitfulData.ktrit

UnitfulData.mHart

UnitfulData.mdibit

UnitfulData.mnat

UnitfulData.mnybl

UnitfulData.msh

UnitfulData.mtribit

UnitfulData.mtrit

UnitfulData.nHart

UnitfulData.nat

UnitfulData.ndibit

UnitfulData.nnat

UnitfulData.nnybl

UnitfulData.nsh

UnitfulData.ntribit

UnitfulData.ntrit

UnitfulData.nybl

UnitfulData.pHart

UnitfulData.pdibit

UnitfulData.pnat

UnitfulData.pnybl

UnitfulData.prefixed_data

UnitfulData.psh

UnitfulData.ptribit

UnitfulData.ptrit

UnitfulData.sh

UnitfulData.tribit

UnitfulData.trit

UnitfulData.yHart

UnitfulData.ydibit

UnitfulData.ynat

UnitfulData.ynybl

UnitfulData.ysh

UnitfulData.ytribit

UnitfulData.ytrit

UnitfulData.zHart

UnitfulData.zdibit

UnitfulData.znat

UnitfulData.znybl

UnitfulData.zsh

UnitfulData.ztribit

UnitfulData.ztrit

UnitfulData.μHart

UnitfulData.μdibit

UnitfulData.μnat

UnitfulData.μnybl

UnitfulData.μsh

UnitfulData.μtribit

UnitfulData.μtrit

UnitfulData.DataRate

UnitfulData.DataRateFreeUnits

UnitfulData.DataRateUnits

UnitfulData.Information

UnitfulData.InformationFreeUnits

UnitfulData.InformationUnits

UnitfulData.data_summary

UnitfulData.@data_allocated

UnitfulData.@unit_custom_prefix

UnitfulData.Bps — Constant

UnitfulData.Bps

The byte per second, a unit of data transfer rate, defined as 1 byte / s.

Dimension: UnitfulData.DataRate

See also: - UnitfulData.DataRate - UnitfulData.bps

UnitfulData.Byte — Constant

UnitfulData.Byte

The Byte, a SI unit of information, defined as 2^3 bits.

Dimension: UnitfulData.Data

See also: - UnitfulData.Data - UnitfulData.bit

UnitfulData.Data — Constant

UnitfulData.Data

A dimension representing information.

UnitfulData.DataRate — Type

UnitfulData.DataRate{T, U}

A supertype for quantities and levels of dimension Data * 𝐓 ^ -1 with a value of type T and units U.

See also: Unitful.Quantity, Unitful.Level.

UnitfulData.EBps — Constant

UnitfulData.EBps

A prefixed unit, equal to 10^18 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.EByte — Constant

UnitfulData.EByte

A prefixed unit, equal to 10^18 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.EHart — Constant

UnitfulData.EHart

A prefixed unit, equal to 10^18 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Ebit — Constant

UnitfulData.Ebit

A prefixed unit, equal to 10^18 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Ebps — Constant

UnitfulData.Ebps

A prefixed unit, equal to 10^18 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Edibit — Constant

UnitfulData.Edibit

A prefixed unit, equal to 10^18 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.EiBps — Constant

UnitfulData.EiBps

A prefixed unit, equal to 2^60 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.EiByte — Constant

UnitfulData.EiByte

A prefixed unit, equal to 2^60 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Eibit — Constant

UnitfulData.Eibit

A prefixed unit, equal to 2^60 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Eibps — Constant

UnitfulData.Eibps

A prefixed unit, equal to 2^60 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Enat — Constant

UnitfulData.Enat

A prefixed unit, equal to 10^18 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Enybl — Constant

UnitfulData.Enybl

A prefixed unit, equal to 10^18 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Esh — Constant

UnitfulData.Esh

A prefixed unit, equal to 10^18 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Etribit — Constant

UnitfulData.Etribit

A prefixed unit, equal to 10^18 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Etrit — Constant

UnitfulData.Etrit

A prefixed unit, equal to 10^18 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.GBps — Constant

UnitfulData.GBps

A prefixed unit, equal to 10^9 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.GByte — Constant

UnitfulData.GByte

A prefixed unit, equal to 10^9 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.GHart — Constant

UnitfulData.GHart

A prefixed unit, equal to 10^9 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Gbit — Constant

UnitfulData.Gbit

A prefixed unit, equal to 10^9 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Gbps — Constant

UnitfulData.Gbps

A prefixed unit, equal to 10^9 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Gdibit — Constant

UnitfulData.Gdibit

A prefixed unit, equal to 10^9 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.GiBps — Constant

UnitfulData.GiBps

A prefixed unit, equal to 2^30 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.GiByte — Constant

UnitfulData.GiByte

A prefixed unit, equal to 2^30 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Gibit — Constant

UnitfulData.Gibit

A prefixed unit, equal to 2^30 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Gibps — Constant

UnitfulData.Gibps

A prefixed unit, equal to 2^30 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Gnat — Constant

UnitfulData.Gnat

A prefixed unit, equal to 10^9 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Gnybl — Constant

UnitfulData.Gnybl

A prefixed unit, equal to 10^9 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Gsh — Constant

UnitfulData.Gsh

A prefixed unit, equal to 10^9 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Gtribit — Constant

UnitfulData.Gtribit

A prefixed unit, equal to 10^9 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Gtrit — Constant

UnitfulData.Gtrit

A prefixed unit, equal to 10^9 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.Hart — Constant

UnitfulData.sh

The hartley, a base 10 logaritmic unit of information entropy defined as log2(10) sh.

Dimension: UnitfulData.Data

See also: - UnitfulData.Data - UnitfulData.bit - UnitfulData.Hart - UnitfulData.nat

UnitfulData.Information — Type

UnitfulData.Information{T, U}

A supertype for quantities and levels of dimension UnitfulData.Data with a value of type T and units U.

See also: UnitfulData.Data, Unitful.Quantity, Unitful.Level.

UnitfulData.KiBps — Constant

UnitfulData.KiBps

A prefixed unit, equal to 2^10 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.KiByte — Constant

UnitfulData.KiByte

A prefixed unit, equal to 2^10 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Kibit — Constant

UnitfulData.Kibit

A prefixed unit, equal to 2^10 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Kibps — Constant

UnitfulData.Kibps

A prefixed unit, equal to 2^10 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.MBps — Constant

UnitfulData.MBps

A prefixed unit, equal to 10^6 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.MByte — Constant

UnitfulData.MByte

A prefixed unit, equal to 10^6 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.MHart — Constant

UnitfulData.MHart

A prefixed unit, equal to 10^6 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Mbit — Constant

UnitfulData.Mbit

A prefixed unit, equal to 10^6 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Mbps — Constant

UnitfulData.Mbps

A prefixed unit, equal to 10^6 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Mdibit — Constant

UnitfulData.Mdibit

A prefixed unit, equal to 10^6 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.MiBps — Constant

UnitfulData.MiBps

A prefixed unit, equal to 2^20 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.MiByte — Constant

UnitfulData.MiByte

A prefixed unit, equal to 2^20 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Mibit — Constant

UnitfulData.Mibit

A prefixed unit, equal to 2^20 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Mibps — Constant

UnitfulData.Mibps

A prefixed unit, equal to 2^20 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Mnat — Constant

UnitfulData.Mnat

A prefixed unit, equal to 10^6 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Mnybl — Constant

UnitfulData.Mnybl

A prefixed unit, equal to 10^6 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Msh — Constant

UnitfulData.Msh

A prefixed unit, equal to 10^6 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Mtribit — Constant

UnitfulData.Mtribit

A prefixed unit, equal to 10^6 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Mtrit — Constant

UnitfulData.Mtrit

A prefixed unit, equal to 10^6 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.PBps — Constant

UnitfulData.PBps

A prefixed unit, equal to 10^15 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.PByte — Constant

UnitfulData.PByte

A prefixed unit, equal to 10^15 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.PHart — Constant

UnitfulData.PHart

A prefixed unit, equal to 10^15 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Pbit — Constant

UnitfulData.Pbit

A prefixed unit, equal to 10^15 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Pbps — Constant

UnitfulData.Pbps

A prefixed unit, equal to 10^15 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Pdibit — Constant

UnitfulData.Pdibit

A prefixed unit, equal to 10^15 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.PiBps — Constant

UnitfulData.PiBps

A prefixed unit, equal to 2^50 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.PiByte — Constant

UnitfulData.PiByte

A prefixed unit, equal to 2^50 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Pibit — Constant

UnitfulData.Pibit

A prefixed unit, equal to 2^50 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Pibps — Constant

UnitfulData.Pibps

A prefixed unit, equal to 2^50 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Pnat — Constant

UnitfulData.Pnat

A prefixed unit, equal to 10^15 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Pnybl — Constant

UnitfulData.Pnybl

A prefixed unit, equal to 10^15 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Psh — Constant

UnitfulData.Psh

A prefixed unit, equal to 10^15 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Ptribit — Constant

UnitfulData.Ptribit

A prefixed unit, equal to 10^15 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Ptrit — Constant

UnitfulData.Ptrit

A prefixed unit, equal to 10^15 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.TBps — Constant

UnitfulData.TBps

A prefixed unit, equal to 10^12 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.TByte — Constant

UnitfulData.TByte

A prefixed unit, equal to 10^12 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.THart — Constant

UnitfulData.THart

A prefixed unit, equal to 10^12 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Tbit — Constant

UnitfulData.Tbit

A prefixed unit, equal to 10^12 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Tbps — Constant

UnitfulData.Tbps

A prefixed unit, equal to 10^12 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Tdibit — Constant

UnitfulData.Tdibit

A prefixed unit, equal to 10^12 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.TiBps — Constant

UnitfulData.TiBps

A prefixed unit, equal to 2^40 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.TiByte — Constant

UnitfulData.TiByte

A prefixed unit, equal to 2^40 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Tibit — Constant

UnitfulData.Tibit

A prefixed unit, equal to 2^40 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Tibps — Constant

UnitfulData.Tibps

A prefixed unit, equal to 2^40 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Tnat — Constant

UnitfulData.Tnat

A prefixed unit, equal to 10^12 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Tnybl — Constant

UnitfulData.Tnybl

A prefixed unit, equal to 10^12 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Tsh — Constant

UnitfulData.Tsh

A prefixed unit, equal to 10^12 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Ttribit — Constant

UnitfulData.Ttribit

A prefixed unit, equal to 10^12 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Ttrit — Constant

UnitfulData.Ttrit

A prefixed unit, equal to 10^12 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.YBps — Constant

UnitfulData.YBps

A prefixed unit, equal to 10^24 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.YByte — Constant

UnitfulData.YByte

A prefixed unit, equal to 10^24 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.YHart — Constant

UnitfulData.YHart

A prefixed unit, equal to 10^24 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Ybit — Constant

UnitfulData.Ybit

A prefixed unit, equal to 10^24 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Ybps — Constant

UnitfulData.Ybps

A prefixed unit, equal to 10^24 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Ydibit — Constant

UnitfulData.Ydibit

A prefixed unit, equal to 10^24 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.YiBps — Constant

UnitfulData.YiBps

A prefixed unit, equal to 2^80 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.YiByte — Constant

UnitfulData.YiByte

A prefixed unit, equal to 2^80 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Yibit — Constant

UnitfulData.Yibit

A prefixed unit, equal to 2^80 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Yibps — Constant

UnitfulData.Yibps

A prefixed unit, equal to 2^80 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Ynat — Constant

UnitfulData.Ynat

A prefixed unit, equal to 10^24 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Ynybl — Constant

UnitfulData.Ynybl

A prefixed unit, equal to 10^24 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Ysh — Constant

UnitfulData.Ysh

A prefixed unit, equal to 10^24 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Ytribit — Constant

UnitfulData.Ytribit

A prefixed unit, equal to 10^24 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Ytrit — Constant

UnitfulData.Ytrit

A prefixed unit, equal to 10^24 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.ZBps — Constant

UnitfulData.ZBps

A prefixed unit, equal to 10^21 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.ZByte — Constant

UnitfulData.ZByte

A prefixed unit, equal to 10^21 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.ZHart — Constant

UnitfulData.ZHart

A prefixed unit, equal to 10^21 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Zbit — Constant

UnitfulData.Zbit

A prefixed unit, equal to 10^21 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Zbps — Constant

UnitfulData.Zbps

A prefixed unit, equal to 10^21 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Zdibit — Constant

UnitfulData.Zdibit

A prefixed unit, equal to 10^21 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.ZiBps — Constant

UnitfulData.ZiBps

A prefixed unit, equal to 2^70 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.ZiByte — Constant

UnitfulData.ZiByte

A prefixed unit, equal to 2^70 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Zibit — Constant

UnitfulData.Zibit

A prefixed unit, equal to 2^70 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Zibps — Constant

UnitfulData.Zibps

A prefixed unit, equal to 2^70 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Znat — Constant

UnitfulData.Znat

A prefixed unit, equal to 10^21 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Znybl — Constant

UnitfulData.Znybl

A prefixed unit, equal to 10^21 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Zsh — Constant

UnitfulData.Zsh

A prefixed unit, equal to 10^21 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Ztribit — Constant

UnitfulData.Ztribit

A prefixed unit, equal to 10^21 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Ztrit — Constant

UnitfulData.Ztrit

A prefixed unit, equal to 10^21 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.aHart — Constant

UnitfulData.aHart

A prefixed unit, equal to 10^-18 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.adibit — Constant

UnitfulData.adibit

A prefixed unit, equal to 10^-18 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.anat — Constant

UnitfulData.anat

A prefixed unit, equal to 10^-18 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.anybl — Constant

UnitfulData.anybl

A prefixed unit, equal to 10^-18 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.ash — Constant

UnitfulData.ash

A prefixed unit, equal to 10^-18 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.atribit — Constant

UnitfulData.atribit

A prefixed unit, equal to 10^-18 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.atrit — Constant

UnitfulData.atrit

A prefixed unit, equal to 10^-18 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.bit — Constant

UnitfulData.bit

The bit, the SI base unit of information.

Dimension: UnitfulData.Data

See also: - UnitfulData.Data - UnitfulData.sh - UnitfulData.Hart - UnitfulData.nat - UnitfulData.Byte

UnitfulData.bps — Constant

UnitfulData.bps

The byte, a unit of data transfer rate, defined as 1 bit / s.

Dimension: UnitfulData.DataRate

See also: - UnitfulData.DataRate - UnitfulData.Bps

UnitfulData.cHart — Constant

UnitfulData.cHart

A prefixed unit, equal to 10^-2 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.cdibit — Constant

UnitfulData.cdibit

A prefixed unit, equal to 10^-2 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.cnat — Constant

UnitfulData.cnat

A prefixed unit, equal to 10^-2 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.cnybl — Constant

UnitfulData.cnybl

A prefixed unit, equal to 10^-2 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.csh — Constant

UnitfulData.csh

A prefixed unit, equal to 10^-2 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ctribit — Constant

UnitfulData.ctribit

A prefixed unit, equal to 10^-2 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ctrit — Constant

UnitfulData.ctrit

A prefixed unit, equal to 10^-2 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.dHart — Constant

UnitfulData.dHart

A prefixed unit, equal to 10^-1 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.daHart — Constant

UnitfulData.daHart

A prefixed unit, equal to 10^1 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.dadibit — Constant

UnitfulData.dadibit

A prefixed unit, equal to 10^1 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.danat — Constant

UnitfulData.danat

A prefixed unit, equal to 10^1 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.danybl — Constant

UnitfulData.danybl

A prefixed unit, equal to 10^1 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.dash — Constant

UnitfulData.dash

A prefixed unit, equal to 10^1 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.datribit — Constant

UnitfulData.datribit

A prefixed unit, equal to 10^1 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.datrit — Constant

UnitfulData.datrit

A prefixed unit, equal to 10^1 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.ddibit — Constant

UnitfulData.ddibit

A prefixed unit, equal to 10^-1 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.dibit — Constant

UnitfulData.dibit

The dibit, a unit of information, defined as 2*bits.

Dimension: UnitfulData.Data

See also: 
    - [`UnitfulData.Data`](@ref)
    - [`UnitfulData.tribit`](@ref)
    - [`UnitfulData.nybl`](@ref)

UnitfulData.dnat — Constant

UnitfulData.dnat

A prefixed unit, equal to 10^-1 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.dnybl — Constant

UnitfulData.dnybl

A prefixed unit, equal to 10^-1 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.dsh — Constant

UnitfulData.dsh

A prefixed unit, equal to 10^-1 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.dtribit — Constant

UnitfulData.dtribit

A prefixed unit, equal to 10^-1 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.dtrit — Constant

UnitfulData.dtrit

A prefixed unit, equal to 10^-1 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.fHart — Constant

UnitfulData.fHart

A prefixed unit, equal to 10^-15 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.fdibit — Constant

UnitfulData.fdibit

A prefixed unit, equal to 10^-15 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.fnat — Constant

UnitfulData.fnat

A prefixed unit, equal to 10^-15 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.fnybl — Constant

UnitfulData.fnybl

A prefixed unit, equal to 10^-15 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.fsh — Constant

UnitfulData.fsh

A prefixed unit, equal to 10^-15 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ftribit — Constant

UnitfulData.ftribit

A prefixed unit, equal to 10^-15 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ftrit — Constant

UnitfulData.ftrit

A prefixed unit, equal to 10^-15 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.hHart — Constant

UnitfulData.hHart

A prefixed unit, equal to 10^2 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.hdibit — Constant

UnitfulData.hdibit

A prefixed unit, equal to 10^2 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.hnat — Constant

UnitfulData.hnat

A prefixed unit, equal to 10^2 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.hnybl — Constant

UnitfulData.hnybl

A prefixed unit, equal to 10^2 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.hsh — Constant

UnitfulData.hsh

A prefixed unit, equal to 10^2 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.htribit — Constant

UnitfulData.htribit

A prefixed unit, equal to 10^2 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.htrit — Constant

UnitfulData.htrit

A prefixed unit, equal to 10^2 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.kBps — Constant

UnitfulData.kBps

A prefixed unit, equal to 10^3 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.kByte — Constant

UnitfulData.kByte

A prefixed unit, equal to 10^3 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.kHart — Constant

UnitfulData.kHart

A prefixed unit, equal to 10^3 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.kbit — Constant

UnitfulData.kbit

A prefixed unit, equal to 10^3 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.kbps — Constant

UnitfulData.kbps

A prefixed unit, equal to 10^3 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.kdibit — Constant

UnitfulData.kdibit

A prefixed unit, equal to 10^3 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.knat — Constant

UnitfulData.knat

A prefixed unit, equal to 10^3 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.knybl — Constant

UnitfulData.knybl

A prefixed unit, equal to 10^3 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.ksh — Constant

UnitfulData.ksh

A prefixed unit, equal to 10^3 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ktribit — Constant

UnitfulData.ktribit

A prefixed unit, equal to 10^3 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ktrit — Constant

UnitfulData.ktrit

A prefixed unit, equal to 10^3 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.mHart — Constant

UnitfulData.mHart

A prefixed unit, equal to 10^-3 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.mdibit — Constant

UnitfulData.mdibit

A prefixed unit, equal to 10^-3 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.mnat — Constant

UnitfulData.mnat

A prefixed unit, equal to 10^-3 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.mnybl — Constant

UnitfulData.mnybl

A prefixed unit, equal to 10^-3 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.msh — Constant

UnitfulData.msh

A prefixed unit, equal to 10^-3 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.mtribit — Constant

UnitfulData.mtribit

A prefixed unit, equal to 10^-3 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.mtrit — Constant

UnitfulData.mtrit

A prefixed unit, equal to 10^-3 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.nHart — Constant

UnitfulData.nHart

A prefixed unit, equal to 10^-9 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.nat — Constant

UnitfulData.sh

The natural unit of information, a natural logaritmic unit of information entropy defined as 1/log(2) sh.

Dimension: UnitfulData.Data

See also: - UnitfulData.Data - UnitfulData.bit - UnitfulData.Hart - UnitfulData.sh

UnitfulData.ndibit — Constant

UnitfulData.ndibit

A prefixed unit, equal to 10^-9 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.nnat — Constant

UnitfulData.nnat

A prefixed unit, equal to 10^-9 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.nnybl — Constant

UnitfulData.nnybl

A prefixed unit, equal to 10^-9 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.nsh — Constant

UnitfulData.nsh

A prefixed unit, equal to 10^-9 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ntribit — Constant

UnitfulData.ntribit

A prefixed unit, equal to 10^-9 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ntrit — Constant

UnitfulData.ntrit

A prefixed unit, equal to 10^-9 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.nybl — Constant

UnitfulData.nybl

The nibble, a unit of information, defined as 1/2 Bytes or 4 bits.

Dimension: UnitfulData.Data

See also: 
    - [`UnitfulData.Data`](@ref)
    - [`UnitfulData.dibit`](@ref)
    - [`UnitfulData.tribit`](@ref)

UnitfulData.pHart — Constant

UnitfulData.pHart

A prefixed unit, equal to 10^-12 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.pdibit — Constant

UnitfulData.pdibit

A prefixed unit, equal to 10^-12 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.pnat — Constant

UnitfulData.pnat

A prefixed unit, equal to 10^-12 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.pnybl — Constant

UnitfulData.pnybl

A prefixed unit, equal to 10^-12 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.prefixed_data — Constant

const prefixed_data

A dictionary that contains the most common prefixes for unit of Information in metric and IEC format.

The keys of the dictionary are tuples of the form (n, b), where n is the exponent and b is the base. The values of the dictionary are strings representing the corresponding prefixes.

UnitfulData.psh — Constant

UnitfulData.psh

A prefixed unit, equal to 10^-12 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ptribit — Constant

UnitfulData.ptribit

A prefixed unit, equal to 10^-12 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ptrit — Constant

UnitfulData.ptrit

A prefixed unit, equal to 10^-12 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.sh — Constant

UnitfulData.sh

The shannon, the SI base unit of information entropy. The maximum entropy od n bits in n shannon.

Dimension: UnitfulData.Data

See also: - UnitfulData.Data - UnitfulData.bit - UnitfulData.Hart - UnitfulData.nat

UnitfulData.tribit — Constant

UnitfulData.tribit

The tribit, a unit of information, defined as 3 bits.

Dimension: UnitfulData.Data

See also: 
    - [`UnitfulData.Data`](@ref)
    - [`UnitfulData.dibit`](@ref)
    - [`UnitfulData.nybl`](@ref)

UnitfulData.trit — Constant

UnitfulData.trit

The trit, a unit of information, defined as log2(3) bits.

Dimension: UnitfulData.Data

See also: 
    - [`UnitfulData.Data`](@ref)
    - [`UnitfulData.dit`](@ref)
    - [`UnitfulData.nybl`](@ref)

UnitfulData.yHart — Constant

UnitfulData.yHart

A prefixed unit, equal to 10^-24 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.ydibit — Constant

UnitfulData.ydibit

A prefixed unit, equal to 10^-24 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.ynat — Constant

UnitfulData.ynat

A prefixed unit, equal to 10^-24 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.ynybl — Constant

UnitfulData.ynybl

A prefixed unit, equal to 10^-24 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.ysh — Constant

UnitfulData.ysh

A prefixed unit, equal to 10^-24 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ytribit — Constant

UnitfulData.ytribit

A prefixed unit, equal to 10^-24 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ytrit — Constant

UnitfulData.ytrit

A prefixed unit, equal to 10^-24 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.zHart — Constant

UnitfulData.zHart

A prefixed unit, equal to 10^-21 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.zdibit — Constant

UnitfulData.zdibit

A prefixed unit, equal to 10^-21 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.znat — Constant

UnitfulData.znat

A prefixed unit, equal to 10^-21 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.znybl — Constant

UnitfulData.znybl

A prefixed unit, equal to 10^-21 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.zsh — Constant

UnitfulData.zsh

A prefixed unit, equal to 10^-21 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ztribit — Constant

UnitfulData.ztribit

A prefixed unit, equal to 10^-21 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ztrit — Constant

UnitfulData.ztrit

A prefixed unit, equal to 10^-21 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.μHart — Constant

UnitfulData.μHart

A prefixed unit, equal to 10^-6 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.μdibit — Constant

UnitfulData.μdibit

A prefixed unit, equal to 10^-6 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.μnat — Constant

UnitfulData.μnat

A prefixed unit, equal to 10^-6 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.μnybl — Constant

UnitfulData.μnybl

A prefixed unit, equal to 10^-6 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.μsh — Constant

UnitfulData.μsh

A prefixed unit, equal to 10^-6 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.μtribit — Constant

UnitfulData.μtribit

A prefixed unit, equal to 10^-6 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.μtrit — Constant

UnitfulData.μtrit

A prefixed unit, equal to 10^-6 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.DataRateFreeUnits — Type

UnitfulData.DataRateFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension Data * 𝐓 ^ -1. Equivalent to Unitful.FreeUnits{U, Data * 𝐓 ^ -1}.

UnitfulData.DataRateUnits — Type

UnitfulData.DataRateUnits{U}

A supertype for units of dimension Data * 𝐓 ^ -1. Equivalent to Unitful.Units{U, Data * 𝐓 ^ -1}.

See also: Unitful.Units.

UnitfulData.InformationFreeUnits — Type

UnitfulData.InformationFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension UnitfulData.Data. Equivalent to Unitful.FreeUnits{U, UnitfulData.Data}.

See also: UnitfulData.Data.

UnitfulData.InformationUnits — Type

UnitfulData.InformationUnits{U}

A supertype for units of dimension UnitfulData.Data. Equivalent to Unitful.Units{U, UnitfulData.Data}.

See also: UnitfulData.Data, Unitful.Units.

UnitfulData.data_summary — Method

data_summary(ex; unit::Unitlike=Byte, exclude=Union{...}, chargeall=Union{...})

Wrapper around Base.summarysize that returns the size of the data in the specified unit. If unit is not a Data unit, it defaults to Byte.

See also: @data_allocated, Base.summarysize

UnitfulData.@data_allocated — Macro

@data_allocated(ex,unit=Byte)

A macro to evaluate an expression, discarding the resulting value, instead returning the total size of the allocated memory in a unit of UnitfulData.Byte. If unit is specified, the result is converted to that unit. By default, the result is in Bytes.

See also Base.@allocations, data_summary

julia> @data_allocated rand(10^6) MByte
8.000080 MBytes

UnitfulData.@unit_custom_prefix — Macro

@unit_custom_prefix(symb,abbr,name, _prefix,autodocs=false)

A macro to create prefixed units. The macro creates a unit with the specified symbol, abbreviation, and name. The unit is prefixed by the values in _prefix. If autodocs is true, the macro will generate a docstring for the unit.

```julia-repl julia> @unit utest "utest" uTests 1bit false utest julia> @unitcustomprefix utest "utest" uTest _prefix=Dict((2,10)=>"hecto")

See also: Unitful.@unit