Functions and Macros

The Functions and Macros exported by the package:

UnitfulData.data_summary — Method

data_summary(ex; unit::Unitlike=Byte, exclude=Union{...}, chargeall=Union{...})

Wrapper around Base.summarysize that returns the size of the data in the specified unit. If unit is not a Data unit, it defaults to Byte.

See also: @data_allocated, Base.summarysize

UnitfulData.@data_allocated — Macro

@data_allocated(ex,unit=Byte)

A macro to evaluate an expression, discarding the resulting value, instead returning the total size of the allocated memory in a unit of UnitfulData.Byte. If unit is specified, the result is converted to that unit. By default, the result is in Bytes.

See also Base.@allocations, data_summary

julia> @data_allocated rand(10^6) MByte
8.000080 MBytes

UnitfulData.@unit_custom_prefix — Macro

@unit_custom_prefix(symb,abbr,name, _prefix,autodocs=false)

A macro to create prefixed units. The macro creates a unit with the specified symbol, abbreviation, and name. The unit is prefixed by the values in _prefix. If autodocs is true, the macro will generate a docstring for the unit.

```jldoctest julia> @unit utest "utest" uTests 1bit false utest

julia> @unitcustomprefix utest "utest" uTest _prefix=Dict((2,10)=>"hecto") hectotest

```See also: [Unitful.@unit`](@extref)

UnitfulData.Bps — Constant

Bps

The byte per second, a unit of data transfer rate, defined as 1 byte / s.

Dimension: DataRate

See also: [DataRate]((@ref) , bps

UnitfulData.Byte — Constant

Byte

The Byte, a SI unit of information, defined as 2^3 bits.

Dimension: Data

See also: [Data]((@ref) , bit

UnitfulData.Data — Constant

Data

A dimension representing information.

UnitfulData.DataRate — Type

UnitfulData.DataRate{T, U}

A supertype for quantities and levels of dimension Data * 𝐓 ^ -1 with a value of type T and units U.

See also: Unitful.Quantity, Unitful.Level.

UnitfulData.EBps — Constant

UnitfulData.EBps

A prefixed unit, equal to 10^18 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.EByte — Constant

UnitfulData.EByte

A prefixed unit, equal to 10^18 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Ebit — Constant

UnitfulData.Ebit

A prefixed unit, equal to 10^18 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Ebps — Constant

UnitfulData.Ebps

A prefixed unit, equal to 10^18 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.EiBps — Constant

UnitfulData.EiBps

A prefixed unit, equal to 2^60 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.EiByte — Constant

UnitfulData.EiByte

A prefixed unit, equal to 2^60 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Eibit — Constant

UnitfulData.Eibit

A prefixed unit, equal to 2^60 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.GBps — Constant

UnitfulData.GBps

A prefixed unit, equal to 10^9 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.GByte — Constant

UnitfulData.GByte

A prefixed unit, equal to 10^9 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Gbit — Constant

UnitfulData.Gbit

A prefixed unit, equal to 10^9 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Gbps — Constant

UnitfulData.Gbps

A prefixed unit, equal to 10^9 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.GiBps — Constant

UnitfulData.GiBps

A prefixed unit, equal to 2^30 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.GiByte — Constant

UnitfulData.GiByte

A prefixed unit, equal to 2^30 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Gibit — Constant

UnitfulData.Gibit

A prefixed unit, equal to 2^30 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Hart — Constant

sh

The hartley, a base 10 logaritmic unit of information entropy defined as log2(10) sh.

Dimension: Data

See also: [Data]((@ref) , [bit]((@ref) , [Hart]((@ref) , nat

UnitfulData.KiBps — Constant

UnitfulData.KiBps

A prefixed unit, equal to 2^10 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.KiByte — Constant

UnitfulData.KiByte

A prefixed unit, equal to 2^10 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Kibit — Constant

UnitfulData.Kibit

A prefixed unit, equal to 2^10 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.MBps — Constant

UnitfulData.MBps

A prefixed unit, equal to 10^6 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.MByte — Constant

UnitfulData.MByte

A prefixed unit, equal to 10^6 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Mbit — Constant

UnitfulData.Mbit

A prefixed unit, equal to 10^6 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Mbps — Constant

UnitfulData.Mbps

A prefixed unit, equal to 10^6 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.MiBps — Constant

UnitfulData.MiBps

A prefixed unit, equal to 2^20 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.MiByte — Constant

UnitfulData.MiByte

A prefixed unit, equal to 2^20 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Mibit — Constant

UnitfulData.Mibit

A prefixed unit, equal to 2^20 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.PBps — Constant

UnitfulData.PBps

A prefixed unit, equal to 10^15 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.PByte — Constant

UnitfulData.PByte

A prefixed unit, equal to 10^15 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Pbit — Constant

UnitfulData.Pbit

A prefixed unit, equal to 10^15 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Pbps — Constant

UnitfulData.Pbps

A prefixed unit, equal to 10^15 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.PiBps — Constant

UnitfulData.PiBps

A prefixed unit, equal to 2^50 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.PiByte — Constant

UnitfulData.PiByte

A prefixed unit, equal to 2^50 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Pibit — Constant

UnitfulData.Pibit

A prefixed unit, equal to 2^50 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.TBps — Constant

UnitfulData.TBps

A prefixed unit, equal to 10^12 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.TByte — Constant

UnitfulData.TByte

A prefixed unit, equal to 10^12 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Tbit — Constant

UnitfulData.Tbit

A prefixed unit, equal to 10^12 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Tbps — Constant

UnitfulData.Tbps

A prefixed unit, equal to 10^12 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.TiBps — Constant

UnitfulData.TiBps

A prefixed unit, equal to 2^40 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.TiByte — Constant

UnitfulData.TiByte

A prefixed unit, equal to 2^40 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Tibit — Constant

UnitfulData.Tibit

A prefixed unit, equal to 2^40 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.YBps — Constant

UnitfulData.YBps

A prefixed unit, equal to 10^24 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.YByte — Constant

UnitfulData.YByte

A prefixed unit, equal to 10^24 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Ybit — Constant

UnitfulData.Ybit

A prefixed unit, equal to 10^24 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Ybps — Constant

UnitfulData.Ybps

A prefixed unit, equal to 10^24 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.YiBps — Constant

UnitfulData.YiBps

A prefixed unit, equal to 2^80 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.YiByte — Constant

UnitfulData.YiByte

A prefixed unit, equal to 2^80 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Yibit — Constant

UnitfulData.Yibit

A prefixed unit, equal to 2^80 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.ZBps — Constant

UnitfulData.ZBps

A prefixed unit, equal to 10^21 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.ZByte — Constant

UnitfulData.ZByte

A prefixed unit, equal to 10^21 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Zbit — Constant

UnitfulData.Zbit

A prefixed unit, equal to 10^21 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.Zbps — Constant

UnitfulData.Zbps

A prefixed unit, equal to 10^21 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.ZiBps — Constant

UnitfulData.ZiBps

A prefixed unit, equal to 2^70 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.ZiByte — Constant

UnitfulData.ZiByte

A prefixed unit, equal to 2^70 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.Zibit — Constant

UnitfulData.Zibit

A prefixed unit, equal to 2^70 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.bit — Constant

bit

The bit, the SI base unit of information.

Dimension: Data

See also: [Data]((@ref) , [sh]((@ref) , [Hart]((@ref) , [nat]((@ref) , Byte

UnitfulData.bps — Constant

bps

The byte, a unit of data transfer rate, defined as 1 bit / s.

Dimension: DataRate

See also: [DataRate]((@ref) , Bps

UnitfulData.dibit — Constant

dibit

The dibit, a unit of information, defined as 2*bits.

Dimension: Data

See also: [`Data`]((@ref) , [`tribit`]((@ref) , [`nybl`](@ref)

UnitfulData.kBps — Constant

UnitfulData.kBps

A prefixed unit, equal to 10^3 Bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.Bps.

UnitfulData.kByte — Constant

UnitfulData.kByte

A prefixed unit, equal to 10^3 Byte.

Dimension: Data

See also: UnitfulData.Byte.

UnitfulData.kbit — Constant

UnitfulData.kbit

A prefixed unit, equal to 10^3 bit.

Dimension: Data

See also: UnitfulData.bit.

UnitfulData.kbps — Constant

UnitfulData.kbps

A prefixed unit, equal to 10^3 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.nat — Constant

sh

The natural unit of information, a natural logaritmic unit of information entropy defined as 1/log(2) sh.

Dimension: Data

See also: [Data]((@ref) , [bit]((@ref) , [Hart]((@ref) , sh

UnitfulData.nybl — Constant

nybl

The nibble, a unit of information, defined as 1/2 Bytes or 4 bits.

Dimension: Data

See also: [`Data`]((@ref) , [`dibit`]((@ref) , [`tribit`](@ref)

UnitfulData.prefixed_data — Constant

const prefixed_data

A dictionary that contains the most common prefixes for unit of Information in metric and IEC format.

The keys of the dictionary are tuples of the form (n, b), where n is the exponent and b is the base. The values of the dictionary are strings representing the corresponding prefixes.

UnitfulData.sh — Constant

sh

The shannon, the SI base unit of information entropy. The maximum entropy od n bits in n shannon.

Dimension: Data

See also: [Data]((@ref) , [bit]((@ref) , [Hart]((@ref) , nat

UnitfulData.tribit — Constant

tribit

The tribit, a unit of information, defined as 3 bits.

Dimension: Data

See also: [`Data`]((@ref) , [`dibit`]((@ref) , [`nybl`](@ref)

UnitfulData.trit — Constant

trit

The trit, a unit of information, defined as log2(3) bits.

Dimension: Data

See also: [`Data`]((@ref) , [`dit`]((@ref) , [`nybl`](@ref)