Units

Exported

UnitfulData.DataRateFreeUnits — Type

UnitfulData.DataRateFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension Data * 𝐓 ^ -1. Equivalent to Unitful.FreeUnits{U, Data * 𝐓 ^ -1}.

UnitfulData.DataRateUnits — Type

UnitfulData.DataRateUnits{U}

A supertype for units of dimension Data * 𝐓 ^ -1. Equivalent to Unitful.Units{U, Data * 𝐓 ^ -1}.

See also: Unitful.Units.

UnitfulData.InformationFreeUnits — Type

UnitfulData.InformationFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension UnitfulData.Data. Equivalent to Unitful.FreeUnits{U, UnitfulData.Data}.

See also: UnitfulData.Data.

UnitfulData.InformationUnits — Type

UnitfulData.InformationUnits{U}

A supertype for units of dimension UnitfulData.Data. Equivalent to Unitful.Units{U, UnitfulData.Data}.

See also: UnitfulData.Data, Unitful.Units.

UnitfulData.prefixed_data — Constant

const prefixed_data

A dictionary that contains the most common prefixes for unit of Information in metric and IEC format.

The keys of the dictionary are tuples of the form (n, b), where n is the exponent and b is the base. The values of the dictionary are strings representing the corresponding prefixes.

Not exported

UnitfulData.EHart — Constant

UnitfulData.EHart

A prefixed unit, equal to 10^18 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Edibit — Constant

UnitfulData.Edibit

A prefixed unit, equal to 10^18 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.Eibps — Constant

UnitfulData.Eibps

A prefixed unit, equal to 2^60 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Enat — Constant

UnitfulData.Enat

A prefixed unit, equal to 10^18 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Enybl — Constant

UnitfulData.Enybl

A prefixed unit, equal to 10^18 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Esh — Constant

UnitfulData.Esh

A prefixed unit, equal to 10^18 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Etribit — Constant

UnitfulData.Etribit

A prefixed unit, equal to 10^18 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Etrit — Constant

UnitfulData.Etrit

A prefixed unit, equal to 10^18 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.GHart — Constant

UnitfulData.GHart

A prefixed unit, equal to 10^9 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Gdibit — Constant

UnitfulData.Gdibit

A prefixed unit, equal to 10^9 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.Gibps — Constant

UnitfulData.Gibps

A prefixed unit, equal to 2^30 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Gnat — Constant

UnitfulData.Gnat

A prefixed unit, equal to 10^9 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Gnybl — Constant

UnitfulData.Gnybl

A prefixed unit, equal to 10^9 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Gsh — Constant

UnitfulData.Gsh

A prefixed unit, equal to 10^9 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Gtribit — Constant

UnitfulData.Gtribit

A prefixed unit, equal to 10^9 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Gtrit — Constant

UnitfulData.Gtrit

A prefixed unit, equal to 10^9 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.Information — Type

UnitfulData.Information{T, U}

A supertype for quantities and levels of dimension UnitfulData.Data with a value of type T and units U.

See also: UnitfulData.Data, Unitful.Quantity, Unitful.Level.

UnitfulData.Kibps — Constant

UnitfulData.Kibps

A prefixed unit, equal to 2^10 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.MHart — Constant

UnitfulData.MHart

A prefixed unit, equal to 10^6 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Mdibit — Constant

UnitfulData.Mdibit

A prefixed unit, equal to 10^6 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.Mibps — Constant

UnitfulData.Mibps

A prefixed unit, equal to 2^20 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Mnat — Constant

UnitfulData.Mnat

A prefixed unit, equal to 10^6 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Mnybl — Constant

UnitfulData.Mnybl

A prefixed unit, equal to 10^6 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Msh — Constant

UnitfulData.Msh

A prefixed unit, equal to 10^6 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Mtribit — Constant

UnitfulData.Mtribit

A prefixed unit, equal to 10^6 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Mtrit — Constant

UnitfulData.Mtrit

A prefixed unit, equal to 10^6 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.PHart — Constant

UnitfulData.PHart

A prefixed unit, equal to 10^15 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Pdibit — Constant

UnitfulData.Pdibit

A prefixed unit, equal to 10^15 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.Pibps — Constant

UnitfulData.Pibps

A prefixed unit, equal to 2^50 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Pnat — Constant

UnitfulData.Pnat

A prefixed unit, equal to 10^15 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Pnybl — Constant

UnitfulData.Pnybl

A prefixed unit, equal to 10^15 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Psh — Constant

UnitfulData.Psh

A prefixed unit, equal to 10^15 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Ptribit — Constant

UnitfulData.Ptribit

A prefixed unit, equal to 10^15 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Ptrit — Constant

UnitfulData.Ptrit

A prefixed unit, equal to 10^15 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.THart — Constant

UnitfulData.THart

A prefixed unit, equal to 10^12 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Tdibit — Constant

UnitfulData.Tdibit

A prefixed unit, equal to 10^12 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.Tibps — Constant

UnitfulData.Tibps

A prefixed unit, equal to 2^40 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Tnat — Constant

UnitfulData.Tnat

A prefixed unit, equal to 10^12 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Tnybl — Constant

UnitfulData.Tnybl

A prefixed unit, equal to 10^12 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Tsh — Constant

UnitfulData.Tsh

A prefixed unit, equal to 10^12 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Ttribit — Constant

UnitfulData.Ttribit

A prefixed unit, equal to 10^12 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Ttrit — Constant

UnitfulData.Ttrit

A prefixed unit, equal to 10^12 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.YHart — Constant

UnitfulData.YHart

A prefixed unit, equal to 10^24 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Ydibit — Constant

UnitfulData.Ydibit

A prefixed unit, equal to 10^24 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.Yibps — Constant

UnitfulData.Yibps

A prefixed unit, equal to 2^80 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Ynat — Constant

UnitfulData.Ynat

A prefixed unit, equal to 10^24 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Ynybl — Constant

UnitfulData.Ynybl

A prefixed unit, equal to 10^24 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Ysh — Constant

UnitfulData.Ysh

A prefixed unit, equal to 10^24 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Ytribit — Constant

UnitfulData.Ytribit

A prefixed unit, equal to 10^24 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Ytrit — Constant

UnitfulData.Ytrit

A prefixed unit, equal to 10^24 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.ZHart — Constant

UnitfulData.ZHart

A prefixed unit, equal to 10^21 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.Zdibit — Constant

UnitfulData.Zdibit

A prefixed unit, equal to 10^21 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.Zibps — Constant

UnitfulData.Zibps

A prefixed unit, equal to 2^70 bps.

Dimension: Data 𝐓^-1

See also: UnitfulData.bps.

UnitfulData.Znat — Constant

UnitfulData.Znat

A prefixed unit, equal to 10^21 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.Znybl — Constant

UnitfulData.Znybl

A prefixed unit, equal to 10^21 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.Zsh — Constant

UnitfulData.Zsh

A prefixed unit, equal to 10^21 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.Ztribit — Constant

UnitfulData.Ztribit

A prefixed unit, equal to 10^21 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.Ztrit — Constant

UnitfulData.Ztrit

A prefixed unit, equal to 10^21 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.aHart — Constant

UnitfulData.aHart

A prefixed unit, equal to 10^-18 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.adibit — Constant

UnitfulData.adibit

A prefixed unit, equal to 10^-18 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.anat — Constant

UnitfulData.anat

A prefixed unit, equal to 10^-18 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.anybl — Constant

UnitfulData.anybl

A prefixed unit, equal to 10^-18 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.ash — Constant

UnitfulData.ash

A prefixed unit, equal to 10^-18 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.atribit — Constant

UnitfulData.atribit

A prefixed unit, equal to 10^-18 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.atrit — Constant

UnitfulData.atrit

A prefixed unit, equal to 10^-18 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.cHart — Constant

UnitfulData.cHart

A prefixed unit, equal to 10^-2 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.cdibit — Constant

UnitfulData.cdibit

A prefixed unit, equal to 10^-2 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.cnat — Constant

UnitfulData.cnat

A prefixed unit, equal to 10^-2 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.cnybl — Constant

UnitfulData.cnybl

A prefixed unit, equal to 10^-2 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.csh — Constant

UnitfulData.csh

A prefixed unit, equal to 10^-2 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ctribit — Constant

UnitfulData.ctribit

A prefixed unit, equal to 10^-2 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ctrit — Constant

UnitfulData.ctrit

A prefixed unit, equal to 10^-2 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.dHart — Constant

UnitfulData.dHart

A prefixed unit, equal to 10^-1 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.daHart — Constant

UnitfulData.daHart

A prefixed unit, equal to 10^1 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.dadibit — Constant

UnitfulData.dadibit

A prefixed unit, equal to 10^1 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.danat — Constant

UnitfulData.danat

A prefixed unit, equal to 10^1 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.danybl — Constant

UnitfulData.danybl

A prefixed unit, equal to 10^1 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.dash — Constant

UnitfulData.dash

A prefixed unit, equal to 10^1 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.datribit — Constant

UnitfulData.datribit

A prefixed unit, equal to 10^1 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.datrit — Constant

UnitfulData.datrit

A prefixed unit, equal to 10^1 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.ddibit — Constant

UnitfulData.ddibit

A prefixed unit, equal to 10^-1 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.dnat — Constant

UnitfulData.dnat

A prefixed unit, equal to 10^-1 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.dnybl — Constant

UnitfulData.dnybl

A prefixed unit, equal to 10^-1 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.dsh — Constant

UnitfulData.dsh

A prefixed unit, equal to 10^-1 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.dtribit — Constant

UnitfulData.dtribit

A prefixed unit, equal to 10^-1 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.dtrit — Constant

UnitfulData.dtrit

A prefixed unit, equal to 10^-1 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.fHart — Constant

UnitfulData.fHart

A prefixed unit, equal to 10^-15 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.fdibit — Constant

UnitfulData.fdibit

A prefixed unit, equal to 10^-15 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.fnat — Constant

UnitfulData.fnat

A prefixed unit, equal to 10^-15 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.fnybl — Constant

UnitfulData.fnybl

A prefixed unit, equal to 10^-15 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.fsh — Constant

UnitfulData.fsh

A prefixed unit, equal to 10^-15 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ftribit — Constant

UnitfulData.ftribit

A prefixed unit, equal to 10^-15 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ftrit — Constant

UnitfulData.ftrit

A prefixed unit, equal to 10^-15 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.hHart — Constant

UnitfulData.hHart

A prefixed unit, equal to 10^2 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.hdibit — Constant

UnitfulData.hdibit

A prefixed unit, equal to 10^2 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.hnat — Constant

UnitfulData.hnat

A prefixed unit, equal to 10^2 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.hnybl — Constant

UnitfulData.hnybl

A prefixed unit, equal to 10^2 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.hsh — Constant

UnitfulData.hsh

A prefixed unit, equal to 10^2 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.htribit — Constant

UnitfulData.htribit

A prefixed unit, equal to 10^2 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.htrit — Constant

UnitfulData.htrit

A prefixed unit, equal to 10^2 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.kHart — Constant

UnitfulData.kHart

A prefixed unit, equal to 10^3 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.kdibit — Constant

UnitfulData.kdibit

A prefixed unit, equal to 10^3 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.knat — Constant

UnitfulData.knat

A prefixed unit, equal to 10^3 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.knybl — Constant

UnitfulData.knybl

A prefixed unit, equal to 10^3 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.ksh — Constant

UnitfulData.ksh

A prefixed unit, equal to 10^3 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ktribit — Constant

UnitfulData.ktribit

A prefixed unit, equal to 10^3 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ktrit — Constant

UnitfulData.ktrit

A prefixed unit, equal to 10^3 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.mHart — Constant

UnitfulData.mHart

A prefixed unit, equal to 10^-3 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.mdibit — Constant

UnitfulData.mdibit

A prefixed unit, equal to 10^-3 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.mnat — Constant

UnitfulData.mnat

A prefixed unit, equal to 10^-3 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.mnybl — Constant

UnitfulData.mnybl

A prefixed unit, equal to 10^-3 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.msh — Constant

UnitfulData.msh

A prefixed unit, equal to 10^-3 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.mtribit — Constant

UnitfulData.mtribit

A prefixed unit, equal to 10^-3 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.mtrit — Constant

UnitfulData.mtrit

A prefixed unit, equal to 10^-3 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.nHart — Constant

UnitfulData.nHart

A prefixed unit, equal to 10^-9 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.ndibit — Constant

UnitfulData.ndibit

A prefixed unit, equal to 10^-9 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.nnat — Constant

UnitfulData.nnat

A prefixed unit, equal to 10^-9 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.nnybl — Constant

UnitfulData.nnybl

A prefixed unit, equal to 10^-9 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.nsh — Constant

UnitfulData.nsh

A prefixed unit, equal to 10^-9 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ntribit — Constant

UnitfulData.ntribit

A prefixed unit, equal to 10^-9 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ntrit — Constant

UnitfulData.ntrit

A prefixed unit, equal to 10^-9 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.pHart — Constant

UnitfulData.pHart

A prefixed unit, equal to 10^-12 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.pdibit — Constant

UnitfulData.pdibit

A prefixed unit, equal to 10^-12 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.pnat — Constant

UnitfulData.pnat

A prefixed unit, equal to 10^-12 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.pnybl — Constant

UnitfulData.pnybl

A prefixed unit, equal to 10^-12 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.psh — Constant

UnitfulData.psh

A prefixed unit, equal to 10^-12 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ptribit — Constant

UnitfulData.ptribit

A prefixed unit, equal to 10^-12 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ptrit — Constant

UnitfulData.ptrit

A prefixed unit, equal to 10^-12 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.yHart — Constant

UnitfulData.yHart

A prefixed unit, equal to 10^-24 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.ydibit — Constant

UnitfulData.ydibit

A prefixed unit, equal to 10^-24 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.ynat — Constant

UnitfulData.ynat

A prefixed unit, equal to 10^-24 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.ynybl — Constant

UnitfulData.ynybl

A prefixed unit, equal to 10^-24 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.ysh — Constant

UnitfulData.ysh

A prefixed unit, equal to 10^-24 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ytribit — Constant

UnitfulData.ytribit

A prefixed unit, equal to 10^-24 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ytrit — Constant

UnitfulData.ytrit

A prefixed unit, equal to 10^-24 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.zHart — Constant

UnitfulData.zHart

A prefixed unit, equal to 10^-21 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.zdibit — Constant

UnitfulData.zdibit

A prefixed unit, equal to 10^-21 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.znat — Constant

UnitfulData.znat

A prefixed unit, equal to 10^-21 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.znybl — Constant

UnitfulData.znybl

A prefixed unit, equal to 10^-21 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.zsh — Constant

UnitfulData.zsh

A prefixed unit, equal to 10^-21 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.ztribit — Constant

UnitfulData.ztribit

A prefixed unit, equal to 10^-21 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.ztrit — Constant

UnitfulData.ztrit

A prefixed unit, equal to 10^-21 trit.

Dimension: Data

See also: UnitfulData.trit.

UnitfulData.μHart — Constant

UnitfulData.μHart

A prefixed unit, equal to 10^-6 Hart.

Dimension: Data

See also: UnitfulData.Hart.

UnitfulData.μdibit — Constant

UnitfulData.μdibit

A prefixed unit, equal to 10^-6 dibit.

Dimension: Data

See also: UnitfulData.dibit.

UnitfulData.μnat — Constant

UnitfulData.μnat

A prefixed unit, equal to 10^-6 nat.

Dimension: Data

See also: UnitfulData.nat.

UnitfulData.μnybl — Constant

UnitfulData.μnybl

A prefixed unit, equal to 10^-6 nybl.

Dimension: Data

See also: UnitfulData.nybl.

UnitfulData.μsh — Constant

UnitfulData.μsh

A prefixed unit, equal to 10^-6 sh.

Dimension: Data

See also: UnitfulData.sh.

UnitfulData.μtribit — Constant

UnitfulData.μtribit

A prefixed unit, equal to 10^-6 tribit.

Dimension: Data

See also: UnitfulData.tribit.

UnitfulData.μtrit — Constant

UnitfulData.μtrit

A prefixed unit, equal to 10^-6 trit.

Dimension: Data

See also: UnitfulData.trit.