Units

The units exported by the package.

UnitfulData.Bps — Constant

Bps

The byte per second, a unit of data transfer rate, defined as 1 byte / s.

Dimension: DataRate

See also: - DataRate - bps

UnitfulData.Byte — Constant

Byte

The Byte, a SI unit of information, defined as 2^3 bits.

Dimension: Data

See also: - Data - bit

UnitfulData.Data — Constant

Data

A dimension representing information.

UnitfulData.Hart — Constant

sh

The hartley, a base 10 logaritmic unit of information entropy defined as log2(10) sh.

Dimension: Data

See also: - Data - bit - Hart - nat

UnitfulData.bit — Constant

bit

The bit, the SI base unit of information.

Dimension: Data

See also: - Data - sh - Hart - nat - Byte

UnitfulData.bps — Constant

bps

The byte, a unit of data transfer rate, defined as 1 bit / s.

Dimension: DataRate

See also: - DataRate - Bps

UnitfulData.dibit — Constant

dibit

The dibit, a unit of information, defined as 2*bits.

Dimension: Data

See also: 
    - [`Data`](@ref)
    - [`tribit`](@ref)
    - [`nybl`](@ref)

UnitfulData.nat — Constant

sh

The natural unit of information, a natural logaritmic unit of information entropy defined as 1/log(2) sh.

Dimension: Data

See also: - Data - bit - Hart - sh

UnitfulData.nybl — Constant

nybl

The nibble, a unit of information, defined as 1/2 Bytes or 4 bits.

Dimension: Data

See also: 
    - [`Data`](@ref)
    - [`dibit`](@ref)
    - [`tribit`](@ref)

UnitfulData.sh — Constant

sh

The shannon, the SI base unit of information entropy. The maximum entropy od n bits in n shannon.

Dimension: Data

See also: - Data - bit - Hart - nat

UnitfulData.tribit — Constant

tribit

The tribit, a unit of information, defined as 3 bits.

Dimension: Data

See also: 
    - [`Data`](@ref)
    - [`dibit`](@ref)
    - [`nybl`](@ref)

UnitfulData.trit — Constant

trit

The trit, a unit of information, defined as log2(3) bits.

Dimension: Data

See also: 
    - [`Data`](@ref)
    - [`dit`](@ref)
    - [`nybl`](@ref)